+++ lib/core/services/services.dart
@@ -6,5 +6,5 @@ library;
 export 'hive_box_manager.dart';
 export 'install_config_service.dart';
 export 'launch_at_startup_service.dart';
-export 'local_auth_failure.dart';
-export 'local_auth_service.dart';
+export 'local_auth_service/local_auth_failure.dart';
+export 'local_auth_service/local_auth_service.dart';