+++ lib/features/password_manager/shared/widgets/custom_fields/widgets/custom_fields_editor.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:hoplixi/core/theme/index.dart';
+import 'package:hoplixi/core/theme/theme.dart';
 import 'package:hoplixi/main_db/core/models/enums/entity_types.dart';
 import 'package:hoplixi/features/password_manager/shared/widgets/custom_fields/models/custom_field_entry.dart';
 import 'package:hoplixi/shared/ui/text_field.dart';